+++ core-js/blocks-filters/responsive-settings/index.tsx
@@ -163,7 +163,7 @@ const initResponsiveSettings = () => {
     attributes: {},
     inlineStyleGenerator,
     classNameGenerator: generateClassNames,
-    order: 'before',
+    order: 'after',
     Edit: BlockEdit,
   });
 };